producer/requirements.txt
kafka-python
pyspark
requests
#finnhub-python